move todolib into todoapp
