--- examples/nav-lib/.size-snapshot.json
@@ -1,36 +1,0 @@
-{
-  "dist/index.umd.production.js": {
-    "bundled": 32702,
-    "minified": 9138,
-    "gzipped": 3236
-  },
-  "dist/index.umd.development.js": {
-    "bundled": 32826,
-    "minified": 9213,
-    "gzipped": 3257
-  },
-  "dist/index.production.js": {
-    "bundled": 3441,
-    "minified": 2155,
-    "gzipped": 834
-  },
-  "dist/index.development.js": {
-    "bundled": 3441,
-    "minified": 2155,
-    "gzipped": 834
-  },
-  "dist/index.es.js": {
-    "bundled": 3288,
-    "minified": 2039,
-    "gzipped": 793,
-    "treeshaked": {
-      "rollup": {
-        "code": 29,
-        "import_statements": 29
-      },
-      "webpack": {
-        "code": 1013
-      }
-    }
-  }
-}
