--- examples/nav-lib/src/index.ts
@@ -1,2 +1,0 @@
-export * from './NavLib';
-export { default } from './NavLib';
