--- examples/nav-lib/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "jsx": "react",
-    "moduleResolution": "node",
-    "outDir": "dist",
-    "target": "ES2017"
-  }
-}
