--- examples/settings-lib/.eslintrc.js
@@ -1,1 +1,0 @@
-module.exports = require('unfig').getCfg(__filename);
