--- examples/settings-lib/.size-snapshot.json
@@ -1,36 +1,0 @@
-{
-  "dist/index.umd.production.js": {
-    "bundled": 33171,
-    "minified": 9318,
-    "gzipped": 3251
-  },
-  "dist/index.umd.development.js": {
-    "bundled": 33295,
-    "minified": 9393,
-    "gzipped": 3271
-  },
-  "dist/index.production.js": {
-    "bundled": 3869,
-    "minified": 2371,
-    "gzipped": 878
-  },
-  "dist/index.development.js": {
-    "bundled": 3869,
-    "minified": 2371,
-    "gzipped": 878
-  },
-  "dist/index.es.js": {
-    "bundled": 3716,
-    "minified": 2255,
-    "gzipped": 837,
-    "treeshaked": {
-      "rollup": {
-        "code": 29,
-        "import_statements": 29
-      },
-      "webpack": {
-        "code": 1013
-      }
-    }
-  }
-}
