--- examples/settings-lib/.unfig.js
@@ -1,3 +1,0 @@
-module.exports = {
-  toolkits: [require('@unfig/toolkit-react-comp')],
-};
