--- examples/settings-lib/src/index.ts
@@ -1,2 +1,0 @@
-export * from './SettingsLib';
-export { default } from './SettingsLib';
