--- examples/todo-lib/.prettierrc.js
@@ -1,1 +1,0 @@
-module.exports = require('unfig').getCfg(__filename);
