--- examples/todo-lib/.size-snapshot.json
@@ -1,36 +1,0 @@
-{
-  "dist/index.umd.production.js": {
-    "bundled": 39982,
-    "minified": 11644,
-    "gzipped": 4128
-  },
-  "dist/index.umd.development.js": {
-    "bundled": 40106,
-    "minified": 11719,
-    "gzipped": 4148
-  },
-  "dist/index.production.js": {
-    "bundled": 3895,
-    "minified": 2212,
-    "gzipped": 793
-  },
-  "dist/index.development.js": {
-    "bundled": 3895,
-    "minified": 2212,
-    "gzipped": 793
-  },
-  "dist/index.es.js": {
-    "bundled": 3665,
-    "minified": 2029,
-    "gzipped": 729,
-    "treeshaked": {
-      "rollup": {
-        "code": 270,
-        "import_statements": 84
-      },
-      "webpack": {
-        "code": 1347
-      }
-    }
-  }
-}
