--- examples/todo-lib/src/index.ts
@@ -1,2 +1,0 @@
-export * from './TodoLib';
-export { default } from './TodoLib';
