+++ examples/todoapp/package.json
@@ -2,10 +2,8 @@
   "name": "@stated-library/todoapp",
   "version": "1.0.6",
   "dependencies": {
-    "@stated-library/nav-lib": "^1.0.6",
+    "@stated-library/core": "^1.0.6",
     "@stated-library/react": "^1.0.6",
-    "@stated-library/settings-lib": "^1.0.6",
-    "@stated-library/todo-lib": "^1.0.6",
     "classnames": "^2.2.6",
     "react": "^16.8.3",
     "react-dom": "^16.8.3",
